--- src/main/java/wooteco/subway/service/favorite/exception/StationNotFoundException.java
@@ -1,7 +1,0 @@
-package wooteco.subway.service.favorite.exception;
-
-public class StationNotFoundException extends FavoriteException {
-	public StationNotFoundException() {
-		super("해당하는 역이 존재하지 않습니다.");
-	}
-}
