+++ src/main/java/wooteco/subway/web/exception/favorite/FavoriteException.java
@@ -1,4 +1,4 @@
-package wooteco.subway.service.favorite.exception;
+package wooteco.subway.web.exception.favorite;
 
 public class FavoriteException extends RuntimeException {
 	public FavoriteException() {
