+++ src/main/java/wooteco/subway/web/exception/member/EmailNotFoundException.java
@@ -1,4 +1,4 @@
-package wooteco.subway.service.member.exception;
+package wooteco.subway.web.exception.member;
 
 public class EmailNotFoundException extends MemberException {
 	public EmailNotFoundException() {
