+++ src/main/java/wooteco/subway/web/exception/member/MemberException.java
@@ -1,4 +1,4 @@
-package wooteco.subway.service.member.exception;
+package wooteco.subway.web.exception.member;
 
 public class MemberException extends RuntimeException {
 	public MemberException() {
