+++ src/main/java/wooteco/subway/web/exception/path/TargetEqualsSourceException.java
@@ -1,0 +1,7 @@
+package wooteco.subway.web.exception.path;
+
+public class TargetEqualsSourceException extends PathException {
+	public TargetEqualsSourceException() {
+		super("출발역과 도착역이 같습니다.");
+	}
+}
