+++ src/main/java/wooteco/subway/web/favorite/ControllerAdvice.java
@@ -1,4 +1,4 @@
-package wooteco.subway.web;
+package wooteco.subway.web.favorite;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.MethodArgumentNotValidException;
@@ -6,8 +6,8 @@ import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 
 @RestControllerAdvice
-public class GeneralControllerAdvice {
-	@ExceptionHandler(MethodArgumentNotValidException.class)
+public class ControllerAdvice {
+	@ExceptionHandler(RuntimeException.class)
 	public ResponseEntity handleMethodArgumentNotValidException(MethodArgumentNotValidException e) {
 		return ResponseEntity.badRequest().body(e.getMessage());
 	}
