+++ src/main/java/wooteco/subway/web/member/LoginMemberMethodArgumentResolver.java
@@ -12,7 +12,7 @@ import org.springframework.web.method.support.ModelAndViewContainer;
 
 import wooteco.subway.domain.member.Member;
 import wooteco.subway.service.member.MemberService;
-import wooteco.subway.service.member.exception.InvalidAuthenticationException;
+import wooteco.subway.web.exception.member.InvalidAuthenticationException;
 
 @Component
 public class LoginMemberMethodArgumentResolver implements HandlerMethodArgumentResolver {
